separate menu into its own file
