+++ src/app/state/modal.ts
@@ -1,7 +1,9 @@
 import { atom } from 'jotai';
 import type { MatrixEvent, Room, Relations } from '$types/matrix-sdk';
+import type { OptionMenuProps } from '$components/message/modals/Options';
 
 export enum ModalType {
+  MobileOptions = 'mobile_options',
   Delete = 'delete',
   // for forwarding a message to another room, not to be confused with the "share" action which is for sharing a message to another app
   Forward = 'forward',
@@ -13,6 +15,7 @@ export enum ModalType {
 }
 
 export type ModalState =
+  | { type: ModalType.MobileOptions; options: OptionMenuProps }
   | { type: ModalType.Delete; room: Room; mEvent: MatrixEvent }
   | { type: ModalType.Forward; room: Room; mEvent: MatrixEvent }
   | { type: ModalType.Report; room: Room; mEvent: MatrixEvent }
